requirements.txt
# Core
pytz==2022.1
Django==4.0.3
django-configurations==2.3.2
gunicorn==20.1.0
newrelic==7.8.0.174

# For the persistence stores
psycopg2-binary==2.9.3
SQLAlchemy==1.4.36 

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.1
Markdown==3.3.6
django-filter==21.1
django-cors-headers==3.13.0

# Developer Tools
django-extensions==3.2.0
ipython==8.2.0
mkdocs==1.3.0
flake8==4.0.1
python-dotenv===0.20.0

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
#nose-progressive==1.5.2
coverage==6.3.2

# Static and Media Storage
django-storages==1.12.3
boto3==1.21.31

# Email
django-ses==3.0.1

# Phonenumber field
django-phonenumber-field==6.1.0
phonenumbers==8.12.46 

# Debugging
sentry-sdk==1.5.10

# Schedualer
APScheduler==3.9.1

# Redis
redis==4.2.2
django-redis==5.2.0

# Utils
PyYAML==6.0
ua-parser==0.16.1
user-agents==2.2.0
django-user-agents==0.4.0
django-tinymce==3.5.0
django-jazzmin==2.6.0
